boule working but not for translation

+++ src/renderer/canvas.c
@@ -6,13 +6,13 @@
 /*   By: rpoder <rpoder@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/19 12:43:59 by rpoder            #+#    #+#             */
-/*   Updated: 2022/10/26 11:51:54 by rpoder           ###   ########.fr       */
+/*   Updated: 2022/10/28 12:02:23 by rpoder           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-void	my_mlx_pixel_put(t_image *image, int x, int y, int color)
+void	my_mlx_pixel_put(t_image *image, int x, int y, long color)
 {
 	char	*dst;
